refactor(glens): extract leaf components (LiveElapsed, CopyButton, MessageFooter, LoadingBubble, UserBubble, AnswerBubble, DashboardBubble, Sidebar, ChatInput)
