+++ apps/web/src/components/glens/GLensChatPage.tsx
@@ -6,598 +6,26 @@ import { useRouter, usePathname, useSearchParams } from "next/navigation"
 import { useAuthFetch } from "@/hooks/useAuthFetch"
 import { useLensEvent } from "@/hooks/useLensEvent"
 import { useLensSessionStream, type LensSessionStream } from "@/hooks/useLensSessionStream"
-import { GlensDashboard } from "@/components/glens/GlensDashboard"
 import type { GlensDashboardSpec } from "@/components/glens/GlensDashboard"
 import { GlensPageBubble } from "@/components/glens/GlensPageBubble"
 import { GenericTableBubble } from "@/components/glens/GenericTableBubble"
 import { BlocksBubble } from "@/components/glens/BlocksBubble"
-import { FeedbackButtons } from "@/components/glens/FeedbackButtons"
 import RunDetailPanel, { type RunMeta } from "@/components/runs/RunDetailPanel"
-import { SlashDropdown, SlashForm, filterTools, type SlashTool } from "@/components/glens/SlashPicker"
 import type { GLensSession, PolicyMapping, MessageBody, Message, RunBlockState } from "@/components/glens/glensTypes"
 import { withId, replaceLast, replaceById } from "@/components/glens/glensTypes"
 import { DEFAULT_SUGGESTIONS, PAGE_SUGGESTIONS, SKILL_LABELS, SKILL_APPLY_URL, _applyBody } from "@/components/glens/glensConstants"
-import { renderMd } from "@/components/glens/glensMarkdown"
-import { formatElapsed } from "@/components/glens/formatElapsed"
+import { Sidebar } from "@/components/glens/Sidebar"
+import { ChatInput } from "@/components/glens/ChatInput"
+import { LiveElapsed } from "@/components/glens/LiveElapsed"
+import { CopyButton } from "@/components/glens/CopyButton"
+import { MessageFooter } from "@/components/glens/MessageFooter"
+import { UserBubble } from "@/components/glens/bubbles/UserBubble"
+import { AnswerBubble } from "@/components/glens/bubbles/AnswerBubble"
+import { LoadingBubble } from "@/components/glens/bubbles/LoadingBubble"
+import { DashboardBubble } from "@/components/glens/bubbles/DashboardBubble"
 
-// ─── Sidebar ──────────────────────────────────────────────────────────────────
-
-function bucketByDay(sessions: GLensSession[]): { label: string; items: GLensSession[] }[] {
-  const startOfDay = (d: Date) => { const c = new Date(d); c.setHours(0, 0, 0, 0); return c.getTime() }
-  const today = startOfDay(new Date())
-  const buckets: Record<string, GLensSession[]> = { today: [], yesterday: [], prev7: [], older: [] }
-  for (const s of sessions) {
-    const diff = Math.floor((today - startOfDay(new Date(s.created_at))) / 86_400_000)
-    if (diff <= 0) buckets.today.push(s)
-    else if (diff === 1) buckets.yesterday.push(s)
-    else if (diff <= 7) buckets.prev7.push(s)
-    else buckets.older.push(s)
-  }
-  return [
-    { label: "Today", items: buckets.today },
-    { label: "Yesterday", items: buckets.yesterday },
-    { label: "Previous 7 days", items: buckets.prev7 },
-    { label: "Older", items: buckets.older },
-  ].filter(b => b.items.length > 0)
-}
-
-function SessionRow({
-  session, active, onSelect, onDelete, onRename,
-}: {
-  session: GLensSession
-  active: boolean
-  onSelect: () => void
-  onDelete: () => void
-  onRename: (title: string) => void
-}) {
-  const [hover, setHover] = useState(false)
-  const [editing, setEditing] = useState(false)
-  const [draft, setDraft] = useState(session.title)
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  useEffect(() => {
-    if (editing && inputRef.current) {
-      inputRef.current.focus()
-      inputRef.current.select()
-    }
-  }, [editing])
-
-  function commit() {
-    const trimmed = draft.trim()
-    if (trimmed && trimmed !== session.title) onRename(trimmed)
-    setEditing(false)
-  }
-
-  function cancel() {
-    setDraft(session.title)
-    setEditing(false)
-  }
-
-  return (
-    <div
-      onMouseEnter={() => setHover(true)}
-      onMouseLeave={() => setHover(false)}
-      style={{ position: "relative", marginBottom: 2 }}
-    >
-      {editing ? (
-        <input
-          ref={inputRef}
-          value={draft}
-          onChange={e => setDraft(e.target.value)}
-          onBlur={commit}
-          onKeyDown={e => {
-            if (e.key === "Enter") { e.preventDefault(); commit() }
-            if (e.key === "Escape") { e.preventDefault(); cancel() }
-          }}
-          style={{
-            width: "100%",
-            padding: "7px 10px",
-            borderRadius: 6,
-            border: "1px solid var(--accent)",
-            background: "var(--surface)",
-            color: "var(--text)",
-            fontSize: 13,
-            outline: "none",
-            fontFamily: "inherit",
-          }}
-        />
-      ) : (
-        <button
-          onClick={onSelect}
-          onDoubleClick={() => setEditing(true)}
-          style={{
-            width: "100%",
-            textAlign: "left",
-            padding: "7px 52px 7px 10px",
-            borderRadius: 6,
-            border: "none",
-            background: active ? "var(--accent-weak)" : hover ? "var(--surface-3, rgba(0,0,0,0.04))" : "transparent",
-            cursor: "pointer",
-            fontSize: 13,
-            color: active ? "var(--accent-text)" : "var(--text)",
-            fontWeight: active ? 600 : 400,
-            overflow: "hidden",
-            textOverflow: "ellipsis",
-            whiteSpace: "nowrap",
-          }}
-        >
-          {session.title}
-        </button>
-      )}
-      {!editing && (
-        <div style={{
-          position: "absolute",
-          right: 4,
-          top: "50%",
-          transform: "translateY(-50%)",
-          display: "flex",
-          gap: 2,
-          opacity: hover ? 1 : 0,
-          transition: "opacity 120ms",
-        }}>
-          <button
-            onClick={() => { setDraft(session.title); setEditing(true) }}
-            aria-label="Rename conversation"
-            title="Rename"
-            style={{
-              padding: "3px 5px",
-              borderRadius: 4,
-              border: "none",
-              background: "transparent",
-              color: "var(--text-muted)",
-              cursor: "pointer",
-              lineHeight: 0,
-            }}
-          >
-            <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-              <path d="M12 20h9" />
-              <path d="M16.5 3.5a2.121 2.121 0 013 3L7 19l-4 1 1-4 12.5-12.5z" />
-            </svg>
-          </button>
-          <button
-            onClick={onDelete}
-            aria-label="Delete conversation"
-            title="Delete"
-            style={{
-              padding: "2px 6px",
-              borderRadius: 4,
-              border: "none",
-              background: "transparent",
-              color: "var(--text-muted)",
-              cursor: "pointer",
-              fontSize: 14,
-              lineHeight: 1,
-            }}
-          >
-            ×
-          </button>
-        </div>
-      )}
-    </div>
-  )
-}
-
-function Sidebar({
-  sessions,
-  activeId,
-  onSelect,
-  onDelete,
-  onRename,
-  onNew,
-  collapsed,
-  onToggle,
-}: {
-  sessions: GLensSession[]
-  activeId: string | null
-  onSelect: (id: string) => void
-  onDelete: (id: string) => void
-  onRename: (id: string, title: string) => void
-  onNew: () => void
-  collapsed: boolean
-  onToggle: () => void
-}) {
-  const grouped = bucketByDay(sessions)
-
-  if (collapsed) {
-    return (
-      <div style={{
-        width: 52,
-        flexShrink: 0,
-        borderRight: "1px solid var(--border)",
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        background: "var(--surface-2)",
-        height: "100%",
-        padding: "12px 0",
-        gap: 8,
-      }}>
-        <button
-          onClick={onToggle}
-          aria-label="Expand sidebar"
-          title="Expand sidebar"
-          style={{
-            width: 32, height: 32, borderRadius: 6, border: "none",
-            background: "transparent", color: "var(--text-muted)",
-            cursor: "pointer", fontSize: 16, lineHeight: 1,
-          }}
-        >
-          ›
-        </button>
-        <button
-          onClick={onNew}
-          aria-label="New chat"
-          title="New chat"
-          style={{
-            width: 32, height: 32, borderRadius: 6,
-            border: "1px solid var(--border)",
-            background: "var(--surface)", color: "var(--text)",
-            cursor: "pointer", fontSize: 16, lineHeight: 1,
-          }}
-        >
-          +
-        </button>
-      </div>
-    )
-  }
-
-  return (
-    <div style={{
-      width: 260,
-      flexShrink: 0,
-      borderRight: "1px solid var(--border)",
-      display: "flex",
-      flexDirection: "column",
-      background: "var(--surface-2)",
-      height: "100%",
-    }}>
-      <div style={{ padding: "14px 12px", borderBottom: "1px solid var(--border)" }}>
-        <div style={{ display: "flex", alignItems: "center", justifyContent: "space-between", marginBottom: 10 }}>
-          <div style={{ fontSize: 12, fontWeight: 700, color: "var(--accent-text)", textTransform: "uppercase", letterSpacing: ".08em" }}>Lens</div>
-          <button
-            onClick={onToggle}
-            aria-label="Collapse sidebar"
-            title="Collapse sidebar"
-            style={{
-              padding: "4px 6px", borderRadius: 6, border: "none",
-              background: "transparent", cursor: "pointer",
-              color: "var(--text-muted)", fontSize: 14, lineHeight: 1,
-            }}
-          >
-            ‹
-          </button>
-        </div>
-        <button
-          onClick={onNew}
-          style={{
-            width: "100%",
-            fontSize: 13,
-            padding: "9px 12px",
-            borderRadius: 8,
-            border: "1px solid var(--border)",
-            background: "var(--surface)",
-            color: "var(--text)",
-            cursor: "pointer",
-            fontWeight: 600,
-            display: "flex",
-            alignItems: "center",
-            gap: 8,
-            justifyContent: "flex-start",
-          }}
-        >
-          <span style={{ fontSize: 15, lineHeight: 1 }}>+</span>
-          <span>New chat</span>
-        </button>
-      </div>
-
-      <div style={{ flex: 1, overflowY: "auto", padding: "10px 8px" }}>
-        {sessions.length === 0 && (
-          <div style={{ fontSize: 12, color: "var(--text-muted)", textAlign: "center", padding: "24px 8px" }}>
-            No conversations yet
-          </div>
-        )}
-        {grouped.map(bucket => (
-          <div key={bucket.label} style={{ marginBottom: 14 }}>
-            <div style={{
-              fontSize: 10,
-              fontWeight: 700,
-              color: "var(--text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: ".08em",
-              padding: "4px 8px 6px",
-            }}>
-              {bucket.label}
-            </div>
-            {bucket.items.map(s => (
-              <SessionRow
-                key={s.id}
-                session={s}
-                active={s.id === activeId}
-                onSelect={() => onSelect(s.id)}
-                onDelete={() => onDelete(s.id)}
-                onRename={(title) => onRename(s.id, title)}
-              />
-            ))}
-          </div>
-        ))}
-      </div>
-
-      <div style={{ padding: "12px 14px", borderTop: "1px solid var(--border)", fontSize: 11, color: "var(--text-muted)", display: "flex", alignItems: "center", justifyContent: "space-between" }}>
-        <span>AI governance analytics</span>
-        <a href="/theguard/team-memory" style={{ fontSize: 11, color: "var(--accent-text)", textDecoration: "none", fontWeight: 500 }}>Team Memory →</a>
-      </div>
-    </div>
-  )
-}
 
 // ─── Message bubbles ──────────────────────────────────────────────────────────
-
-function UserBubble({ text, onEdit }: { text: string; onEdit?: (newText: string) => void }) {
-  const [editing, setEditing] = useState(false)
-  const [draft, setDraft] = useState(text)
-  const taRef = useRef<HTMLTextAreaElement>(null)
-
-  function startEdit() { setDraft(text); setEditing(true); setTimeout(() => taRef.current?.focus(), 0) }
-  function cancel() { setEditing(false) }
-  function submit() { if (draft.trim() && draft.trim() !== text) onEdit?.(draft.trim()); setEditing(false) }
-
-  return (
-    <div
-      style={{ display: "flex", justifyContent: "flex-end", marginBottom: 16, position: "relative" }}
-      onMouseEnter={e => { if (!editing && onEdit) (e.currentTarget.querySelector(".edit-btn") as HTMLElement)?.style.setProperty("opacity", "1") }}
-      onMouseLeave={e => { (e.currentTarget.querySelector(".edit-btn") as HTMLElement)?.style.setProperty("opacity", "0") }}
-    >
-      {onEdit && !editing && (
-        <button
-          className="edit-btn"
-          onClick={startEdit}
-          style={{
-            opacity: 0, transition: "opacity .15s", alignSelf: "center", marginRight: 8,
-            background: "none", border: "none", cursor: "pointer", color: "var(--text-muted)",
-            fontSize: 13, padding: "2px 6px", borderRadius: 4,
-          }}
-          title="Edit question"
-        >✎</button>
-      )}
-      {editing ? (
-        <div style={{ maxWidth: "65%", display: "flex", flexDirection: "column", gap: 6 }}>
-          <textarea
-            ref={taRef}
-            value={draft}
-            onChange={e => setDraft(e.target.value)}
-            onKeyDown={e => { if (e.key === "Enter" && !e.shiftKey) { e.preventDefault(); submit() } if (e.key === "Escape") cancel() }}
-            rows={3}
-            style={{
-              width: "100%", fontSize: 14, padding: "10px 14px", borderRadius: 10,
-              border: "1px solid var(--accent)", outline: "none", resize: "none",
-              background: "var(--surface-2)", color: "var(--text)", lineHeight: 1.5,
-            }}
-          />
-          <div style={{ display: "flex", gap: 6, justifyContent: "flex-end" }}>
-            <button onClick={cancel} style={{ fontSize: 12, padding: "4px 10px", borderRadius: 6, border: "1px solid var(--border)", background: "var(--surface-2)", cursor: "pointer", color: "var(--text-2)" }}>Cancel</button>
-            <button onClick={submit} style={{ fontSize: 12, padding: "4px 10px", borderRadius: 6, border: "none", background: "var(--accent)", color: "#fff", cursor: "pointer" }}>Send</button>
-          </div>
-        </div>
-      ) : (
-        <div style={{
-          maxWidth: "65%",
-          background: "var(--accent)",
-          color: "#fff",
-          borderRadius: "14px 14px 4px 14px",
-          padding: "10px 16px",
-          fontSize: 14,
-          lineHeight: 1.5,
-        }}>
-          {text}
-        </div>
-      )}
-    </div>
-  )
-}
-
-function AnswerBubble({ text, skill, drilldown, followups, onFollowup, understoodAs }: { text: string; skill?: string; drilldown?: { path: string }; followups?: string[]; onFollowup?: (q: string) => void; understoodAs?: string }) {
-  return (
-    <div style={{ display: "flex", justifyContent: "flex-start", marginBottom: 16 }}>
-      <div style={{ maxWidth: "75%" }}>
-        <div style={{ display: "flex", alignItems: "center", gap: 8, marginBottom: 4 }}>
-          {skill && (
-            <div style={{ fontSize: 10, fontWeight: 700, color: "var(--text-muted)", textTransform: "uppercase", letterSpacing: ".06em" }}>
-              {SKILL_LABELS[skill] ?? skill}
-            </div>
-          )}
-          {understoodAs && (
-            <div style={{ fontSize: 10, color: "var(--text-muted)", fontStyle: "italic" }}>
-              · {understoodAs}
-            </div>
-          )}
-        </div>
-        <div style={{
-          background: "var(--surface-2)",
-          border: "1px solid var(--border)",
-          borderRadius: "4px 14px 14px 14px",
-          padding: "10px 16px",
-          fontSize: 14,
-          color: "var(--text)",
-          lineHeight: 1.6,
-        }}>
-          {renderMd(text)}
-          {drilldown && (
-            <div style={{ marginTop: 8, textAlign: "right" }}>
-              <a href={drilldown.path} style={{ fontSize: 12, color: "var(--accent, #6366f1)", textDecoration: "none", fontWeight: 500 }}>
-                View full &rarr;
-              </a>
-            </div>
-          )}
-        </div>
-        {followups && followups.length > 0 && onFollowup && (
-          <div style={{ display: "flex", flexWrap: "wrap", gap: 6, marginTop: 8 }}>
-            {followups.map(q => (
-              <button
-                key={q}
-                onClick={() => onFollowup(q)}
-                style={{
-                  fontSize: 12, padding: "5px 12px", borderRadius: 16,
-                  border: "1px solid var(--border)", background: "var(--surface-2)",
-                  color: "var(--text-2)", cursor: "pointer",
-                }}
-              >
-                {q}
-              </button>
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  )
-}
-
-function LoadingBubble({ label }: { label?: string }) {
-  return (
-    <div style={{ display: "flex", justifyContent: "flex-start", marginBottom: 16 }}>
-      <div style={{
-        background: "var(--surface-2)",
-        border: "1px solid var(--border)",
-        borderRadius: "4px 14px 14px 14px",
-        padding: "12px 16px",
-        fontSize: 14,
-        color: "var(--text-muted)",
-        display: "flex",
-        alignItems: "center",
-        gap: 10,
-      }} aria-label="Lens is thinking" role="status">
-        <span style={{ display: "inline-flex", alignItems: "flex-end", gap: 4, height: 12 }}>
-          <span className="conduct-typing-dot" />
-          <span className="conduct-typing-dot" />
-          <span className="conduct-typing-dot" />
-        </span>
-        {label && <span style={{ fontSize: 12, opacity: 0.8 }}>{label}</span>}
-      </div>
-    </div>
-  )
-}
-
-// ── Copy button (shared) ──────────────────────────────────────────────────────
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false)
-
-  const handleCopy = async () => {
-    try {
-      await navigator.clipboard.writeText(text)
-      setCopied(true)
-      window.setTimeout(() => setCopied(false), 1500)
-    } catch {
-      // Older browsers / insecure contexts — fall back to a hidden textarea.
-      const ta = document.createElement("textarea")
-      ta.value = text
-      ta.style.position = "fixed"
-      ta.style.opacity = "0"
-      document.body.appendChild(ta)
-      ta.select()
-      try { document.execCommand("copy") } catch {}
-      document.body.removeChild(ta)
-      setCopied(true)
-      window.setTimeout(() => setCopied(false), 1500)
-    }
-  }
-
-  return (
-    <button
-      type="button"
-      onClick={handleCopy}
-      aria-label={copied ? "Copied" : "Copy response"}
-      title={copied ? "Copied" : "Copy"}
-      style={{
-        display: "inline-flex",
-        alignItems: "center",
-        justifyContent: "center",
-        width: 26,
-        height: 26,
-        color: copied ? "var(--ok, #10b981)" : "var(--text-muted)",
-        background: "transparent",
-        border: "1px solid transparent",
-        borderRadius: 6,
-        cursor: "pointer",
-        transition: "background 0.12s, color 0.12s",
-      }}
-      onMouseEnter={e => (e.currentTarget.style.background = "var(--surface-3, rgba(0,0,0,0.04))")}
-      onMouseLeave={e => (e.currentTarget.style.background = "transparent")}
-    >
-      {copied ? (
-        <svg width={14} height={14} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round">
-          <polyline points="20 6 9 17 4 12" />
-        </svg>
-      ) : (
-        <svg width={14} height={14} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-          <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
-          <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
-        </svg>
-      )}
-    </button>
-  )
-}
-
-// ── Message footer (copy + thumbs) ────────────────────────────────────────────
-// Rendered after every assistant bubble (except loading + in-flight streaming).
-// Universal: past-session restore, live answers, structured bubbles all get
-// the same affordance in the same place. Feedback requires a sessionId; copy
-// only requires text.
-
-function MessageFooter({ text, sessionId, messageId }: { text?: string; sessionId: string | null; messageId: string }) {
-  if (!text && !sessionId) return null
-  return (
-    <div
-      style={{
-        display: "flex",
-        justifyContent: "flex-start",
-        alignItems: "center",
-        gap: 0,
-        marginTop: 4,
-        marginBottom: 12,
-        marginLeft: -4,
-      }}
-    >
-      {text ? <CopyButton text={text} /> : null}
-      {sessionId ? <FeedbackButtons sessionId={sessionId} messageId={messageId} /> : null}
-    </div>
-  )
-}
-
-function DashboardBubble({
-  spec,
-  sessionId,
-  authFetch,
-  drilldown,
-}: {
-  spec: GlensDashboardSpec
-  sessionId: string
-  authFetch: (url: string, options?: RequestInit) => Promise<Response>
-  drilldown?: { path: string; filters?: Record<string, string> }
-}) {
-  return (
-    <div style={{ display: "flex", justifyContent: "flex-start", marginBottom: 16, width: "100%" }}>
-      <div style={{
-        width: "100%",
-        background: "var(--surface)",
-        border: "1px solid var(--border)",
-        borderRadius: "4px 14px 14px 14px",
-        padding: "16px",
-      }}>
-        <GlensDashboard spec={spec} sessionId={sessionId} authFetch={authFetch} />
-        {drilldown && (
-          <div style={{ marginTop: 12, textAlign: "right" }}>
-            <a
-              href={drilldown.path}
-              style={{ fontSize: 12, color: "var(--accent, #6366f1)", textDecoration: "none", fontWeight: 500 }}
-            >
-              View full &rarr;
-            </a>
-          </div>
-        )}
-      </div>
-    </div>
-  )
-}
 
 function PolicyConfirmBubble({
   answer,
@@ -1478,137 +906,6 @@ function RunBubble({
           </div>
         )}
       </div>
-    </div>
-  )
-}
-
-// Owns its own 1s tick so the elapsed badge updates without re-rendering
-// RunBubble (and, more importantly, the embedded RunDetailPanel below it).
-// Same pattern as <LiveDuration/> inside RunDetailPanel.
-function LiveElapsed({ startedAt, completedAt, status }: {
-  startedAt: number; completedAt: number | null; status: string
-}) {
-  const [now, setNow] = useState<number>(() => Date.now())
-  useEffect(() => {
-    if (completedAt || (status !== "running" && status !== "pending")) return
-    const id = setInterval(() => setNow(Date.now()), 1000)
-    return () => clearInterval(id)
-  }, [completedAt, status])
-  return (
-    <span style={{ fontSize: 11, color: "var(--text-muted)" }}>
-      {formatElapsed(startedAt, completedAt ?? now)}
-    </span>
-  )
-}
-
-// ─── Input ────────────────────────────────────────────────────────────────────
-
-function ChatInput({ onSubmit, disabled }: { onSubmit: (t: string) => void; disabled: boolean }) {
-  const [value, setValue] = useState("")
-  const [pickedTool, setPickedTool] = useState<SlashTool | null>(null)
-  // Escape sets this true to hide the picker without wiping the user's text;
-  // clears when the value stops starting with "/" (fresh slate for next try).
-  const [pickerDismissed, setPickerDismissed] = useState(false)
-  const ref = useRef<HTMLTextAreaElement>(null)
-
-  useEffect(() => { if (!disabled && !pickedTool) ref.current?.focus() }, [disabled, pickedTool])
-  useEffect(() => { if (!value.startsWith("/")) setPickerDismissed(false) }, [value])
-
-  function submit() {
-    const t = value.trim()
-    if (t && !disabled) { onSubmit(t); setValue("") }
-  }
-
-  // Slash-command picker (#1630): only mount the dropdown when there are
-  // real matches. Prevents an invisible dropdown from swallowing Enter for
-  // messages that legitimately start with "/" (e.g. "/tmp/foo is broken").
-  const slashMatches: SlashTool[] =
-    value.startsWith("/") && !pickedTool && !pickerDismissed
-      ? filterTools(value.slice(1))
-      : []
-  const showPicker = slashMatches.length > 0
-
-  if (pickedTool) {
-    return (
-      <SlashForm
-        tool={pickedTool}
-        disabled={disabled}
-        onSubmit={prompt => { onSubmit(prompt); setValue(""); setPickedTool(null) }}
-        onCancel={() => { setPickedTool(null); setValue("") }}
-      />
-    )
-  }
-
-  const canSend = !disabled && value.trim().length > 0
-  return (
-    <div style={{
-      position: "relative",
-      border: "1px solid var(--border)",
-      borderRadius: 16,
-      background: "var(--surface-2)",
-      padding: "10px 14px",
-      transition: "border-color 120ms, box-shadow 120ms",
-    }}>
-      {showPicker && (
-        <SlashDropdown
-          matches={slashMatches}
-          onSelect={t => { setPickedTool(t); setValue("") }}
-          onClose={() => setPickerDismissed(true)}
-        />
-      )}
-      <textarea
-        ref={ref}
-        value={value}
-        onChange={e => setValue(e.target.value)}
-        onKeyDown={e => {
-          if (showPicker) return  // dropdown owns keys only while visible
-          if (e.key === "Enter" && !e.shiftKey) { e.preventDefault(); submit() }
-        }}
-        disabled={disabled}
-        placeholder="Ask about your governance data… (type / for commands)"
-        rows={2}
-        style={{
-          width: "100%",
-          resize: "none",
-          border: "none",
-          padding: 0,
-          paddingRight: 44,
-          fontSize: 14,
-          background: "transparent",
-          color: "var(--text)",
-          outline: "none",
-          fontFamily: "inherit",
-          lineHeight: 1.5,
-          display: "block",
-        }}
-      />
-      <button
-        onClick={submit}
-        disabled={!canSend}
-        aria-label="Send"
-        style={{
-          position: "absolute",
-          right: 8,
-          bottom: 8,
-          width: 32,
-          height: 32,
-          borderRadius: "50%",
-          border: "none",
-          background: canSend ? "var(--text)" : "var(--surface-3, rgba(0,0,0,0.08))",
-          color: canSend ? "var(--surface)" : "var(--text-muted)",
-          cursor: canSend ? "pointer" : "not-allowed",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          transition: "background 120ms, color 120ms",
-          padding: 0,
-        }}
-      >
-        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.4" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
-          <path d="M12 19V5" />
-          <path d="M5 12l7-7 7 7" />
-        </svg>
-      </button>
     </div>
   )
 }
